feat: dependency visualization (#12)
* Add tabbed navigation for result display

Introduced new styles and a tabbed interface for displaying scan results. The tabs allow for future expansion with additional content areas. A JavaScript function ensures the correct tab is shown after the scan request completes.

* Add package dependencies visualization

Introduced a new "Package Dependencies" tab in the UI with a dropdown for selecting packages and a force-directed graph visualization of dependencies. Stored metrics data globally to facilitate this functionality and updated the existing scripts to generate interactive visuals using D3.js.

* Refactor dropdown and search functionality

Enhanced the appearance and structure of dropdown elements using new wrappers and CSS styling. Improved the search functionality for class names by adding a hierarchical highlighting mechanism for better visibility of matching nodes and their connections.

* Add loading spinner to scan form

Implemented a loading spinner to indicate progress during the scan form submission. The spinner appears when the request starts and hides once the request completes, enhancing user experience by providing visual feedback.

* Add dependency visualization section

Updated the README to include a new "Dependency Visualization" section. Added a screenshot to illustrate this feature and highlighted its inclusion in the list of features.